fix(codegen): split gc_map tests to a sibling file for the 2000-line cap
#8948 took gc_map.rs to 2047 lines and `check_file_size` has been red on
main since it merged. I had prepared this split during that PR's review but
pushed it to the wrong remote — #8948 is same-repo, and the push went to a
fork the PR does not track — so the unfixed head is what merged.

Moved the cfg(test) module out verbatim via the #[path] pattern the tree
already uses (1332 + 712 lines).
